refactor(p): add "ContactInfo" as 3. value object
I have separated phone, email and address of patient in this value object. Aggregate root "Patient" now also has a value object "ContactInfo".
